Update nonces to long; move models extensions to separate project

+++ Centaurus.Models/Ledger/LedgerUpdateNotification.cs
@@ -11,7 +11,7 @@ namespace Centaurus.Models
     {
         public override MessageTypes MessageType => MessageTypes.LedgerUpdateNotification;
 
-        public override ulong MessageId => (ulong)LedgerFrom;
+        public override long MessageId => LedgerFrom;
 
         /// <summary>
         /// Ledgers range start.
